bootloader: change pkgstart offset type from int to size_t
Primarily to silence the MSC warnings about conversion from size_t
to int, but size_t also makes more sense as the storage type.

This requires an ugly explict cast to long with fseek() calls to
avoid another warning, but we should actually consider replacing
the fseek() calls with fseeko() (or _fseeki64() under MSC).

+++ bootloader/src/pyi_archive.c
@@ -150,7 +150,7 @@ pyi_arch_extract(ARCHIVE_STATUS *status, TOC *ptoc)
         return NULL;
     }
 
-    fseek(status->fp, status->pkgstart + ntohl(ptoc->pos), SEEK_SET);
+    fseek(status->fp, (long)(status->pkgstart + ntohl(ptoc->pos)), SEEK_SET);
     data = (unsigned char *)malloc(ntohl(ptoc->len));
 
     if (data == NULL) {
@@ -418,7 +418,7 @@ pyi_arch_open(ARCHIVE_STATUS *status)
     pyvers = pyi_arch_get_pyversion(status);
 
     /* Read in in the table of contents */
-    fseek(status->fp, status->pkgstart + ntohl(status->cookie.TOC), SEEK_SET);
+    fseek(status->fp, (long)(status->pkgstart + ntohl(status->cookie.TOC)), SEEK_SET);
     status->tocbuff = (TOC *) malloc(ntohl(status->cookie.TOClen));
 
     if (status->tocbuff == NULL) {
